MDEV-30103 FK constraint fails upon referenced table eviction
ER_NO_REFERENCED_ROW_2 was returned by a piece of code that handled
potentially missing table. In could be missing for example because it was
removed by DROP TABLE under FOREIGN_KEYS_CHECK=OFF.

Here, it happened because table was evicted in lock_release under
innodb_evict_tables_on_commit_debug=1. So yes, it is a debug-only scenario.
By design FK tables were never evicted, but it's changed for debug builds
beginning from 10.3.

2ca11234 already makes some approach to atomically handle table reference
loading and locking.

This is not enough, as table may have changed, or evicted before lock.
In that case, lock_table returns DB_DEADLOCK.

The ultimate solution is seen by me as following:
1. Try calling lock_table. It'll acquire lock_sys, so either acquire the
   lock, or fk_table will be NULL
2. If fk_table is NULL, the error returned was DEADLOCK. Try to call
   dict_table_load.
3. If succeeds, we'll have n_ref_count > 0. Call lock_table again.

Note on double-checked locking:
> Unguarded access to foreign_table/referenced_table may lead to a race
with table load and eviction, if out-of-order execution is considered.
The protection is using a correct acquire/release memory order:

> In lock_fk_table, first `table` argument load matches acquire semantics.
The last load matches release semantics.
In lock_table, load also made to match acquire semantics, just in case.

> In dict_foreign_add_to_cache, load's have release semantics. This
function is used in dict_table_load_one, and foreigns are loaded in the
very end of table initialization.

> In dict_sys::remove, store's are vice-versa having acquire semantics,
because a referential table link is NULLed in the very beginning, which
is also correct. memory_order_acquire can't be applied to
std::atomic::store(), so memory_order_seq_cst is used there.

> For the rest of the places it should be enough to leave relaxed memory
access order, since it is operated under some lock (lock_sys/dict_sys/mdl).

Additional changes:
* lock_fk_table: this new function is dedicated for solving this problem.
Currently, it will be called where lock_table was before, from
row_ins_check_foreign_constraint. This is suboptimal, since this means that
it'll be called for each changed row, and therefore global lock_sys locked.
The call can later be moved to a better place, to make it once per
transaction.

* row_ins_check_foreign_constraints: remove own dict_table_open_on_name
call, as it is now superseded by lock_fk_table.

* dict_table_load: a part of dict_table_open_on_name is extracted into a
new function dedicated to load and check a newly loaded table under
an already locked dict_sys.